+++ package.json
@@ -1,19 +1,57 @@
 {
   "name": "@craftcms/ckeditor",
   "private": true,
-  "main": "gulpfile.js",
+  "main": "webpack.config.js",
+  "browserslist": [
+    "extends @craftcms/browserslist-config"
+  ],
+  "dependencies": {
+    "@ckeditor/ckeditor5-alignment": "^37.0.1",
+    "@ckeditor/ckeditor5-autoformat": "^37.0.1",
+    "@ckeditor/ckeditor5-basic-styles": "^37.0.1",
+    "@ckeditor/ckeditor5-block-quote": "^37.0.1",
+    "@ckeditor/ckeditor5-code-block": "^37.0.1",
+    "@ckeditor/ckeditor5-dev-utils": "^36.0.1",
+    "@ckeditor/ckeditor5-editor-classic": "^37.0.1",
+    "@ckeditor/ckeditor5-essentials": "^37.0.1",
+    "@ckeditor/ckeditor5-find-and-replace": "^37.0.1",
+    "@ckeditor/ckeditor5-heading": "^37.0.1",
+    "@ckeditor/ckeditor5-horizontal-line": "^37.0.1",
+    "@ckeditor/ckeditor5-html-embed": "^37.0.1",
+    "@ckeditor/ckeditor5-html-support": "^37.0.1",
+    "@ckeditor/ckeditor5-image": "^37.0.1",
+    "@ckeditor/ckeditor5-indent": "^37.0.1",
+    "@ckeditor/ckeditor5-inspector": "^4.1.0",
+    "@ckeditor/ckeditor5-link": "^37.0.1",
+    "@ckeditor/ckeditor5-list": "^37.0.1",
+    "@ckeditor/ckeditor5-media-embed": "^37.0.1",
+    "@ckeditor/ckeditor5-page-break": "^37.0.1",
+    "@ckeditor/ckeditor5-paragraph": "^37.0.1",
+    "@ckeditor/ckeditor5-paste-from-office": "^37.0.1",
+    "@ckeditor/ckeditor5-select-all": "^37.0.1",
+    "@ckeditor/ckeditor5-source-editing": "^37.0.1",
+    "@ckeditor/ckeditor5-style": "^37.0.1",
+    "@ckeditor/ckeditor5-table": "^37.0.1",
+    "@ckeditor/ckeditor5-theme-lark": "^37.0.1",
+    "@ckeditor/ckeditor5-word-count": "^37.0.1"
+  },
   "devDependencies": {
-    "craftcms-sass": "^3.5.6",
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^7.0.1",
-    "gulp-clean-css": "^3.10.0",
-    "gulp-sass": "^3.1.0",
-    "gulp-sourcemaps": "^2.6.5",
+    "@craftcms/webpack": "^1.1.0",
+    "@craftcms/sass": "^1.1.0",
+    "css-loader": "^6.7.3",
+    "postcss-loader": "^7.2.4",
+    "raw-loader": "^4.0.2",
+    "style-loader": "^3.3.2",
     "husky": "^7.0.4",
     "lint-staged": "^12.4.0",
     "prettier": "^2.7.1"
   },
   "scripts": {
+    "prebuild": "npm run fix-prettier",
+    "build": "webpack --node-env=production --progress",
+    "serve": "webpack-dev-server --node-env=development",
+    "check-prettier": "prettier --check .",
+    "fix-prettier": "prettier --write .",
     "prepare": "husky install"
   }
 }
