+++ src/Field.php
@@ -4,7 +4,7 @@ namespace craft\ckeditor;
 
 use Craft;
 use craft\base\ElementInterface;
-use craft\ckeditor\assets\ckeditor\CkeditorAsset;
+use craft\ckeditor\web\assets\ckeditor\CkeditorAsset;
 use craft\ckeditor\events\DefineLinkOptionsEvent;
 use craft\elements\Asset;
 use craft\elements\Category;
