--- src/assets/ckeditor/src/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@craftcms/ckeditor5-custom-build",
-  "author": "",
-  "description": "",
-  "license": "ISC",
-  "version": "1.0.0",
-  "main": "index.js",
-  "devDependencies": {
-    "@ckeditor/ckeditor5-alignment": "^37.0.1",
-    "@ckeditor/ckeditor5-autoformat": "^37.0.1",
-    "@ckeditor/ckeditor5-basic-styles": "^37.0.1",
-    "@ckeditor/ckeditor5-block-quote": "^37.0.1",
-    "@ckeditor/ckeditor5-code-block": "^37.0.1",
-    "@ckeditor/ckeditor5-dev-utils": "^36.0.1",
-    "@ckeditor/ckeditor5-editor-classic": "^37.0.1",
-    "@ckeditor/ckeditor5-essentials": "^37.0.1",
-    "@ckeditor/ckeditor5-find-and-replace": "^37.0.1",
-    "@ckeditor/ckeditor5-heading": "^37.0.1",
-    "@ckeditor/ckeditor5-horizontal-line": "^37.0.1",
-    "@ckeditor/ckeditor5-html-embed": "^37.0.1",
-    "@ckeditor/ckeditor5-html-support": "^37.0.1",
-    "@ckeditor/ckeditor5-image": "^37.0.1",
-    "@ckeditor/ckeditor5-indent": "^37.0.1",
-    "@ckeditor/ckeditor5-inspector": "^4.1.0",
-    "@ckeditor/ckeditor5-link": "^37.0.1",
-    "@ckeditor/ckeditor5-list": "^37.0.1",
-    "@ckeditor/ckeditor5-media-embed": "^37.0.1",
-    "@ckeditor/ckeditor5-page-break": "^37.0.1",
-    "@ckeditor/ckeditor5-paragraph": "^37.0.1",
-    "@ckeditor/ckeditor5-paste-from-office": "^37.0.1",
-    "@ckeditor/ckeditor5-select-all": "^37.0.1",
-    "@ckeditor/ckeditor5-source-editing": "^37.0.1",
-    "@ckeditor/ckeditor5-style": "^37.0.1",
-    "@ckeditor/ckeditor5-table": "^37.0.1",
-    "@ckeditor/ckeditor5-theme-lark": "^37.0.1",
-    "@ckeditor/ckeditor5-word-count": "^37.0.1",
-    "css-loader": "^6.7.3",
-    "postcss-loader": "^7.2.4",
-    "raw-loader": "^4.0.2",
-    "style-loader": "^3.3.2",
-    "webpack": "^5.78.0",
-    "webpack-cli": "^5.0.1"
-  },
-  "scripts": {
-    "build": "webpack --mode production",
-    "dev": "webpack --mode development"
-  }
-}
