+++ src/web/assets/ckeditor/CkeditorAsset.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license MIT
+ */
+
+namespace craft\ckeditor\web\assets\ckeditor;
+
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+
+/**
+ * CKEditor custom build asset bundle
+ *
+ * @since 3.0.0
+ */
+class CkeditorAsset extends AssetBundle
+{
+    /**
+     * @inheritdoc
+     */
+    public $sourcePath = __DIR__ . '/dist';
+
+    /**
+     * @inheritdoc
+     */
+    public $depends = [
+        CpAsset::class,
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public $js = [
+        'ckeditor.js',
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public $css = [
+        'ckeditor.css',
+    ];
+}
