+++ webpack.config.js
@@ -1,0 +1,3 @@
+const {getConfigs} = require('@craftcms/webpack');
+
+module.exports = getConfigs();
